+++ database/migrations/2022_11_09_155533_create_concomitans_table.php
@@ -18,22 +18,22 @@ return new class extends Migration
         Schema::create('concomitans', function (Blueprint $table) {
             $table->id();
             $table->foreignIdFor(NurseDoc::class);
-            $table->tinyInteger('a')->nullable()->comment('Гипертоническая болезнь');
-            $table->tinyInteger('b')->nullable()->comment('Ишемическая болезнь сердца. ');
-            $table->tinyInteger('c')->nullable()->comment('В анамнезе перенесенный ИМ');
-            $table->tinyInteger('d')->nullable()->comment('В анамнезе перенесенный ОНМК, ТИА');
-            $table->tinyInteger('e')->nullable()->comment('ХСН');
-            $table->tinyInteger('f')->nullable()->comment('Атеросклероз периферический многососудистый со стенозом и/или эндоартерэктомии в анамнезе, аневризма аорты');
+            $table->float('a', 16,4)->nullable()->comment('Гипертоническая болезнь');
+            $table->float('b', 16,4)->nullable()->comment('Ишемическая болезнь сердца. ');
+            $table->float('c', 16,4)->nullable()->comment('В анамнезе перенесенный ИМ');
+            $table->float('d', 16,4)->nullable()->comment('В анамнезе перенесенный ОНМК, ТИА');
+            $table->float('e', 16,4)->nullable()->comment('ХСН');
+            $table->float('f', 16,4)->nullable()->comment('Атеросклероз периферический многососудистый со стенозом и/или эндоартерэктомии в анамнезе, аневризма аорты');
             $table->enum('g',['1','2','3'])->nullable()->comment('Перенесенные операции на сердце и сосудах');
             $table->enum('h',['1','2','3','4','5','6','7'])->nullable()->comment('Нарушения ритма:');
-            $table->tinyInteger('i')->nullable()->comment('Сахарный диабет без осложнений');
-            $table->tinyInteger('k')->nullable()->comment('Сахарный диабет с осложнениями');
-            $table->tinyInteger('l')->nullable()->comment('Нарушение толерантности к глюкозе');
-            $table->tinyInteger('m')->nullable()->comment('ХБП ');
-            $table->tinyInteger('n')->nullable()->comment('ХОБЛ или БА');
-            $table->tinyInteger('o')->nullable()->comment('Ковид-19');
-            $table->tinyInteger('p')->nullable()->comment('Врожденные и приобретенные пороки сердца');
-            $table->tinyInteger('q')->nullable()->comment('Онкологические заболевания');
+            $table->float('i', 16,4)->nullable()->comment('Сахарный диабет без осложнений');
+            $table->float('k', 16,4)->nullable()->comment('Сахарный диабет с осложнениями');
+            $table->float('l', 16,4)->nullable()->comment('Нарушение толерантности к глюкозе');
+            $table->float('m', 16,4)->nullable()->comment('ХБП ');
+            $table->float('n', 16,4)->nullable()->comment('ХОБЛ или БА');
+            $table->float('o', 16,4)->nullable()->comment('Ковид-19');
+            $table->float('p', 16,4)->nullable()->comment('Врожденные и приобретенные пороки сердца');
+            $table->float('q', 16,4)->nullable()->comment('Онкологические заболевания');
             $table->timestamps();
         });
     }
